refactor(gateway): /api/capabilities sai de health.rs para capabilities.rs
O Quality Ratchet (report-only) contou uma regressao de files_over_700
que e deste PR: health.rs foi de 646 para 739 linhas com FeatureInputs,
feature_flags e o teste de contrato do Garra Mobile. O bloco inteiro de
/api/capabilities (CapabilitiesResponse, FeatureInputs, feature_flags,
capabilities_handler e feature_flags_keep_the_mobile_contract) vira o
modulo capabilities.rs (210 linhas); health.rs volta a 545. Sem mudanca
de comportamento: o router aponta para crate::capabilities e as
referencias em docs/README/ADR/TODO/models.dart seguem o caminho novo.

cargo fmt, cargo check, cargo test (capabilities:: + health::) e
cargo clippy --all-targets -D warnings limpos no crate.
